+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 **Not yet released**
 
+- Added support for dynamic SRV-based backend server discovery as a reverse proxy.
 - Added support for Unix socket backends in forwarded authentication module.
 - Fixed body replacement MIME type filtering logic not being effective at all.
 - Fixed some errors logged two times.